+++ libzen-compositor/callback.h
@@ -1,5 +1,5 @@
-#ifndef LIBZEN_CALLBACK_H
-#define LIBZEN_CALLBACK_H
+#ifndef LIBZEN_COMPOSIOR_CALLBACK_H
+#define LIBZEN_COMPOSIOR_CALLBACK_H
 
 #include <wayland-server.h>
 
@@ -12,4 +12,4 @@ struct zen_callback *zen_callback_create(struct wl_client *client, uint32_t id);
 
 void zen_callback_done(struct zen_callback *callback, uint32_t callback_data);
 
-#endif  //  LIBZEN_CALLBACK_H
+#endif  //  LIBZEN_COMPOSIOR_CALLBACK_H
