+++ zen-backend/headless/output.c
@@ -1,4 +1,4 @@
-#include <libzen/libzen.h>
+#include <libzen-compositor/libzen-compositor.h>
 #include <sys/time.h>
 #include <wayland-server.h>
