+++ zen-renderer/opengl/opengl-component.c
