+++ zen-renderer/opengl/opengl.h
@@ -1,7 +1,7 @@
-#ifndef LIBZEN_OPENGL_H
-#define LIBZEN_OPENGL_H
+#ifndef ZEN_RENDERER_OPENGL_H
+#define ZEN_RENDERER_OPENGL_H
 
-#include <libzen/libzen.h>
+#include <libzen-compositor/libzen-compositor.h>
 
 struct zen_opengl {
   struct zen_compositor* compositor;
@@ -12,4 +12,4 @@ struct zen_opengl* zen_opengl_create(struct zen_compositor* compositor);
 
 void zen_opengl_destroy(struct zen_opengl* opengl);
 
-#endif  //  LIBZEN_OPENGL_H
+#endif  //  ZEN_RENDERER_OPENGL_H
