+++ zen/meson.build
@@ -1,5 +1,5 @@
 dep_zen = [
-  dep_libzen_public,
+  dep_libzen_compositor_public,
   dep_glfw_backend_public, # TODO: load different backend dynamically
 ]
